requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
pymysql==1.1.0
cryptography==41.0.7
